+++ R/rsc_card.R
@@ -53,7 +53,7 @@ rsc_card <- function(content) {
 #' @noRd
 widget_html.rsc_card <- function(id, style, class, ...) {
   htmltools::tagList(
-    # Necessary for RStudio viewer version < 1.2
+    # Necessary for Posit viewer version < 1.2
     reactR::html_dependency_corejs(),
     reactR::html_dependency_react(),
     reactR::html_dependency_reacttools(),