+++ NEWS.md
@@ -5,7 +5,7 @@
 # connectwidgets 0.1.1
 
 * Initial public release and CRAN submission of the {connectwidgets} R package
-* `connect` and `content` functions for enumerating content from RStudio Connect
+* `connect` and `content` functions for enumerating content from Posit Connect
 * `rsc_table`, `rsc_grid`, and `rsc_card` HTML widgets for displaying content
   items in an RMarkdown document
 * `rsc_search` and `rsc_filter` HTML widgets for searching and filtering those